# gin-admin-server/docs/swagger.yaml
definitions:
  request.ChangePassword:
    properties:
      newPassword:
        description: 新密码
        type: string
      oldPassword:
        description: 旧密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
  request.Login:
    properties:
      captcha:
        description: 验证码
        type: string
      captchaId:
        description: 验证码ID
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
  request.Register:
    properties:
      email:
        description: 电子邮箱
        type: string
      headerImg:
        description: 头像链接
        type: string
      nickName:
        description: 昵称
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 联系电话
        type: string
      username:
        description: 用户名
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/c/decodeJwt:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 20000,"data":{},"msg":"解码成功"}'
          schema:
            type: string
      summary: jwt解码,反馈用户信息
      tags:
      - SysJwt
  /api/v1/c/isValidJwt:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 20000,"data":{},"msg":"验证通过"}'
          schema:
            type: string
      summary: jwt验证
      tags:
      - SysJwt
  /api/v1/c/jwt:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 20000,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      summary: 删除用户token
      tags:
      - SysJwt
  /api/v1/sysCaptcha:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 20000,"data":{},"msg":"验证码获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 生成验证码
      tags:
      - SysCaptcha
  /api/v1/sysUser/changePassword:
    post:
      parameters:
      - description: 用户名, 旧密码, 新密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 20000,"data": {},"msg": "修改密码成功"}'
          schema:
            type: string
      summary: 修改密码
      tags:
      - SysUser
  /api/v1/sysUser/login:
    post:
      parameters:
      - description: 用户名, 密码, 验证码，验证码ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 20000,"data":{},"msg":"登陆成功"}'
          schema:
            type: string
      summary: 用户登录
      tags:
      - SysUser
  /api/v1/sysUser/register:
    post:
      parameters:
      - description: 必填：用户名, 密码, 昵称；非必填：头像链接，电子邮箱，联系电话
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Register'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 20000,"data": {},"msg": "注册成功"}'
          schema:
            type: string
      summary: 用户注册
      tags:
      - SysUser
swagger: "2.0"